+++ app/models/user.rb
@@ -3,8 +3,9 @@ class User < ApplicationRecord
   # :confirmable, :lockable, :timeoutable and :omniauthable
   devise :masqueradable, :database_authenticatable, :registerable, :recoverable, :rememberable, :validatable, :omniauthable
 
+  has_one_attached :avatar
   has_person_name
 
-  has_many :notifications, foreign_key: :recipient_id
+  has_many :notifications, as: :recipient
   has_many :services
 end