+++ config/initializers/gravatar.rb
@@ -1,4 +1,0 @@
-GravatarImageTag.configure do |config|
-  config.default_image = "mm"
-  config.secure        = true
-end