+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 2020-08-20
+
+* Add tests for generating postgres and mysql apps
+
+### 2020-08-07
+
+* Refactor notifications to use the [Noticed gem](https://github.com/excid3/noticed)
+
 ### 2019-02-28
 
 * Adds support for Rails 6.0